# docs/swagger.yaml
basePath: /api
definitions:
...
  pipeline.PipelineConfig:
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Enterprise-grade webhook routing and data transformation platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Webhook Router API
  version: "1.0"
paths:
...
  /pipelines:
    get:
      description: Returns a list of all configured data transformation pipelines
      produces:
      - application/json
      responses:
        "200":
          description: List of pipelines
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Pipeline engine not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Get all data pipelines
      tags:
      - pipelines
    post:
      consumes:
      - application/json
      description: Creates a new data transformation pipeline with stages
      parameters:
      - description: Pipeline configuration
        in: body
        name: pipeline
        required: true
        schema:
          $ref: '#/definitions/pipeline.PipelineConfig'
      produces:
      - application/json
      responses:
        "201":
          description: Created pipeline
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON or pipeline validation failed
          schema:
            type: string
        "500":
          description: Failed to register pipeline
          schema:
            type: string
        "503":
          description: Pipeline engine not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Create data pipeline
      tags:
      - pipelines
  /pipelines/{id}:
    delete:
      description: Removes a data transformation pipeline
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Pipeline not found
          schema:
            type: string
        "503":
          description: Pipeline engine not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Delete data pipeline
      tags:
      - pipelines
    get:
      description: Returns a specific data transformation pipeline by ID
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pipeline details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Pipeline not found
          schema:
            type: string
        "503":
          description: Pipeline engine not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Get data pipeline
      tags:
      - pipelines
  /pipelines/{id}/execute:
    post:
      consumes:
      - application/json
      description: Executes a data transformation pipeline with test data
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Test data for pipeline execution
        in: body
        name: testData
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Pipeline execution result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Pipeline not found
          schema:
            type: string
        "500":
          description: Pipeline execution failed
          schema:
            type: string
        "503":
          description: Pipeline engine not initialized
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Execute pipeline
      tags:
      - pipelines
...
  /pipelines/stages/types:
    get:
      description: Returns a list of supported pipeline stage types
      produces:
      - application/json
      responses:
        "200":
          description: List of stage types
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionAuth: []
      summary: Get stage types
      tags:
      - pipelines
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"